Rename Main to App, add short description of pattern and comments for its usage
